+++ Sources/SmockMacro/Generator/VerifierGenerator.swift
@@ -1,149 +1,7 @@
 import SwiftSyntax
 import SwiftSyntaxBuilder
 
-enum FunctionPropertiesGenerator {
-    static func expectationsOptionsClassDeclaration(
-        variablePrefix: String,
-        functionSignature: FunctionSignatureSyntax
-    ) throws -> ClassDeclSyntax {
-
-        var genericParameterClauseElements: [String] = []
-        if functionSignature.effectSpecifiers?.throwsClause?.throwsSpecifier != nil {
-            genericParameterClauseElements.append("ErrorableFieldOptionsProtocol")
-        }
-
-        if functionSignature.returnClause?.type != nil {
-            genericParameterClauseElements.append("ReturnableFieldOptionsProtocol")
-        } else {
-            genericParameterClauseElements.append("VoidReturnableFieldOptionsProtocol")
-        }
-
-        return try ClassDeclSyntax(
-            modifiers: [DeclModifierSyntax(name: "public")],
-            name: "\(raw: variablePrefix.capitalizingComponentsFirstLetter())_FieldOptions",
-            genericParameterClause: genericParameterClauseElements.count > 0
-                ? ": \(raw: genericParameterClauseElements.joined(separator: ", ")) " : nil,
-            memberBlockBuilder: {
-                try VariableDeclSyntax(
-                    """
-                    var expectedResponse: \(raw: variablePrefix.capitalizingComponentsFirstLetter())_ExpectedResponse?
-                    """
-                )
-
-                try VariableDeclSyntax(
-                    """
-                    var times: Int?
-                    """
-                )
-
-                try FunctionDeclSyntax(
-                    """
-                    public func update(using closure: @Sendable @escaping \(ClosureGenerator.closureElements(functionSignature: functionSignature))) {
-                      self.expectedResponse = .closure(closure)
-                    }
-                    """
-                )
-
-                if functionSignature.effectSpecifiers?.throwsClause?.throwsSpecifier != nil {
-                    try FunctionDeclSyntax(
-                        """
-                        public func update(error: Swift.Error) {
-                          self.expectedResponse = .error(error)
-                        }
-                        """
-                    )
-                }
-
-                if let returnType = functionSignature.returnClause?.type {
-                    try FunctionDeclSyntax(
-                        """
-                        public func update(value: \(returnType)) {
-                          self.expectedResponse = .value(value)
-                        }
-                        """
-                    )
-                } else {
-                    try FunctionDeclSyntax(
-                        """
-                        public func success() {
-                          self.expectedResponse = .success
-                        }
-                        """
-                    )
-                }
-
-                try FunctionDeclSyntax(
-                    """
-                    public func update(times: Int?) {
-                      self.times = times
-                    }
-                    """
-                )
-            }
-        )
-    }
-
-    static func expectedResponseEnumDeclaration(
-        variablePrefix: String,
-        functionSignature: FunctionSignatureSyntax
-    ) throws -> EnumDeclSyntax {
-        try EnumDeclSyntax(
-            modifiers: [DeclModifierSyntax(name: "public")],
-            name: "\(raw: variablePrefix.capitalizingComponentsFirstLetter())_ExpectedResponse",
-            genericParameterClause: ": Sendable",
-            memberBlockBuilder: {
-                try EnumCaseDeclSyntax(
-                    """
-                    case closure(@Sendable \(ClosureGenerator.closureElements(functionSignature: functionSignature)))
-                    """
-                )
-
-                if functionSignature.effectSpecifiers?.throwsClause?.throwsSpecifier != nil {
-                    try EnumCaseDeclSyntax(
-                        """
-                        case error(Swift.Error)
-                        """
-                    )
-                }
-
-                if let returnType = functionSignature.returnClause?.type {
-                    try EnumCaseDeclSyntax(
-                        """
-                        case value(\(returnType))
-                        """
-                    )
-                } else {
-                    try EnumCaseDeclSyntax(
-                        """
-                        case success
-                        """
-                    )
-                }
-            }
-        )
-    }
-
-    static func expectedResponseVariableDeclaration(
-        variablePrefix: String,
-        functionDeclaration: FunctionDeclSyntax,
-        accessModifier: String,
-        staticName: Bool
-    ) throws -> VariableDeclSyntax {
-        let expectedResponseType =
-            "\(variablePrefix.capitalizingComponentsFirstLetter())_ExpectedResponse"
-        let variablePrefix = VariablePrefixGenerator.text(for: functionDeclaration)
-        let parameterList = functionDeclaration.signature.parameterClause.parameters
-        let inputMatcherType =
-            parameterList.count > 0
-            ? "\(variablePrefix.capitalizingComponentsFirstLetter())_InputMatcher" : "AlwaysMatcher"
-
-        return try VariableDeclSyntax(
-            """
-            \(raw: accessModifier)var \(raw: staticName ? "expectedResponses" : variablePrefix): [(Int?,\(raw: expectedResponseType),\(raw: inputMatcherType))] = []
-            """
-        )
-    }
-
+enum VerifierGenerator {
     static func verifierStructDeclaration(
         functionDeclarations: [FunctionDeclSyntax],
         isComparableProvider: (String) -> Bool
@@ -295,6 +153,41 @@ enum FunctionPropertiesGenerator {
         return (methodParameters, methodInterpolationParameters, matcherInitializers)
     }
 
+    private static func getWithLockCall(variablePrefix: String) -> FunctionCallExprSyntax {
+        let lockProtectedStatements = CodeBlockItemListSyntax([
+            CodeBlockItemSyntax(
+                item: .stmt(
+                    StmtSyntax(
+                        ReturnStmtSyntax(
+                            expression: ExprSyntax("storage.receivedInvocations.\(raw: variablePrefix)")
+                        )
+                    )
+                )
+            )
+        ])
+
+        let lockClosure = ClosureExprSyntax(
+            signature: ClosureSignatureSyntax(
+                parameterClause: .simpleInput(
+                    ClosureShorthandParameterListSyntax([
+                        ClosureShorthandParameterSyntax(name: .identifier("storage"))
+                    ])
+                )
+            ),
+            statements: lockProtectedStatements
+        )
+
+        return FunctionCallExprSyntax(
+            calledExpression: MemberAccessExprSyntax(
+                base: DeclReferenceExprSyntax(baseName: .identifier("self.state.mutex")),
+                declName: DeclReferenceExprSyntax(baseName: .identifier("withLock"))
+            ),
+            arguments: LabeledExprListSyntax([
+                LabeledExprSyntax(expression: ExprSyntax(lockClosure))
+            ])
+        )
+    }
+
     /// Generate a specific method for a parameter type combination
     private static func generateMethodForCombination(
         functionDeclaration: FunctionDeclSyntax,
@@ -340,39 +233,6 @@ enum FunctionPropertiesGenerator {
             )
         }
 
-        let lockProtectedStatements = CodeBlockItemListSyntax([
-            CodeBlockItemSyntax(
-                item: .stmt(
-                    StmtSyntax(
-                        ReturnStmtSyntax(
-                            expression: ExprSyntax("storage.receivedInvocations.\(raw: variablePrefix)")
-                        )
-                    )
-                )
-            )
-        ])
-
-        let lockClosure = ClosureExprSyntax(
-            signature: ClosureSignatureSyntax(
-                parameterClause: .simpleInput(
-                    ClosureShorthandParameterListSyntax([
-                        ClosureShorthandParameterSyntax(name: .identifier("storage"))
-                    ])
-                )
-            ),
-            statements: lockProtectedStatements
-        )
-
-        let withLockCall = FunctionCallExprSyntax(
-            calledExpression: MemberAccessExprSyntax(
-                base: DeclReferenceExprSyntax(baseName: .identifier("self.state.mutex")),
-                declName: DeclReferenceExprSyntax(baseName: .identifier("withLock"))
-            ),
-            arguments: LabeledExprListSyntax([
-                LabeledExprSyntax(expression: ExprSyntax(lockClosure))
-            ])
-        )
-
         return try FunctionDeclSyntax("public func \(raw: functionName)(\(raw: methodSignature))") {
             VariableDeclSyntax(
                 bindingSpecifier: .keyword(.let),
@@ -381,7 +241,7 @@ enum FunctionPropertiesGenerator {
                         pattern: IdentifierPatternSyntax(identifier: .identifier("invocations")),
                         initializer: InitializerClauseSyntax(
                             equal: .equalToken(),
-                            value: ExprSyntax(withLockCall)
+                            value: ExprSyntax(getWithLockCall(variablePrefix: variablePrefix))
                         )
                     )
                 ])
